quant attention

+++ GPT2_exp.py
@@ -262,7 +262,7 @@ def main():
         eps=1e-8
     )
 
-    accelerator = Accelerator(mixed_precision='fp16')
+    accelerator = Accelerator(mixed_precision='fp8')
     model, optimizer, train_dataloader, eval_dataloader = accelerator.prepare(
         model, optimizer, train_dataloader, eval_dataloader
     )
